Generalize haldane example to arbitrary number of blocks.
